requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow >= 7.0.0
joblib >= 0.14.1
matplotlib >= 3.1.2
numpy >= 1.17.5
opencv-python >= 4.2.0.34
prefetch_generator >= 1.0.1
pyvips >= 2.1.8
scikit_learn >= 0.22.1
scipy >= 1.4.1
torch >= 1.4.0
torchvision >= 0.5.0
tqdm >= 4.42.0
wandb >= 0.8.24
